requirements.txt
streamlit
opencv-python
pytesseract
Pillow
numpy
requests
google-generativeai
